:bug: fix get movie id from navigation in Int type

+++ domain/src/main/kotlin/com/vaniala/movies/domain/usecase/GetMovieDetailsUseCase.kt
@@ -4,5 +4,5 @@ import com.vaniala.movies.domain.repository.MovieRepository
 import javax.inject.Inject
 
 class GetMovieDetailsUseCase @Inject constructor(private val repository: MovieRepository) {
-    operator fun invoke(moveId: Long) = repository.getMovieDetails(moveId)
+    operator fun invoke(moveId: Int) = repository.getMovieDetails(moveId)
 }
